fix screenshare context instance
